# jit: decouple compiled code from moduleInstance's pointer