--- a/internal/wasm/jit/arch_arm64.s
+++ b/internal/wasm/jit/arch_arm64.s
@@ -1,15 +1,17 @@
 #include "funcdata.h"
 #include "textflag.h"
 
-// jitcall(codeSegment, ce)
-TEXT ·jitcall(SB),NOSPLIT|NOFRAME,$0-16
-        // Load the address of native code.
-        MOVD codeSegment+0(FP),R1
-        // Load the address of *callEngine.
+// jitcall(codeSegment, ce, moduleInstanceAddress)
+TEXT ·jitcall(SB),NOSPLIT|NOFRAME,$0-24
+        // Load the address of *callEngine into arm64ReservedRegisterForCallEngine.
         MOVD ce+8(FP),R0
         // In arm64, return address is stored in R30 after jumping into the code.
         // We save the return address value into archContext.jitReturnAddress in Engine.
         // Note that the const 120 drifts after editting Engine or archContext struct. See TestArchContextOffsetInEngine.
         MOVD R30,120(R0)
+        // Load the address of *wasm.ModuleInstance into arm64CallingConvensionModuleInstanceAddressRegister.
+        MOVD moduleInstanceAddress+16(FP),R29
+        // Load the address of native code.
+        MOVD codeSegment+0(FP),R1
         // Jump to native code.
         JMP (R1)