… internal/wasm/jit/jit_controlflow_test.go
… internal/wasm/jit/jit_impl_arm64.go
… internal/wasm/jit/jit_memory_test.go
… internal/wasm/jit/jit_test.go