This commit decouples the *wasm.ModuleInstance from the compiled code.

This is necessary to reuse same machine code for multiple module instances
originating from the same binary (or text), which is a necessary mechanism for
compilation caches I will do in the next PRs.