@mathetake 2022-04-12
ok this doesn't affect the performance:

```
>> benchstat old.txt new.txt
name                                          old time/op    new time/op    delta
Engines/jit/base64_5_per_exec-32                0.00ns ±25%    0.00ns ±24%   ~     (p=0.063 n=5+5)
Engines/jit/base64_100_per_exec-32              0.00ns ±22%    0.00ns ±12%   ~     (p=0.690 n=5+5)
Engines/jit/base64_10000_per_exec-32            0.03ns ± 9%    0.03ns ±15%   ~     (p=0.690 n=5+5)
Engines/jit/fib_for_5-32                         642ns ± 5%     653ns ± 1%   ~     (p=0.310 n=5+5)
Engines/jit/fib_for_10-32                       1.60µs ± 4%    1.65µs ± 4%   ~     (p=0.095 n=5+5)
Engines/jit/fib_for_20-32                       76.8µs ± 2%    79.3µs ± 1%   ~     (p=0.056 n=5+5)
Engines/jit/fib_for_30-32                       8.90ms ± 5%    9.18ms ± 2%   ~     (p=0.151 n=5+5)
Engines/jit/string_manipulation_size_50-32      51.4µs ± 2%    52.2µs ± 2%   ~     (p=0.056 n=5+5)
Engines/jit/string_manipulation_size_100-32      152µs ± 1%     153µs ± 2%   ~     (p=0.690 n=5+5)
Engines/jit/string_manipulation_size_1000-32    10.8ms ± 2%    11.0ms ± 4%   ~     (p=0.222 n=5+5)
Engines/jit/reverse_array_size_500-32           12.1µs ± 2%    12.2µs ± 1%   ~     (p=0.421 n=5+5)
Engines/jit/reverse_array_size_1000-32          23.0µs ± 1%    23.3µs ± 3%   ~     (p=0.310 n=5+5)
Engines/jit/reverse_array_size_10000-32          275µs ± 3%     280µs ± 2%   ~     (p=0.310 n=5+5)
Engines/jit/random_mat_mul_size_5-32            19.0µs ± 1%    18.7µs ± 2%   ~     (p=0.056 n=5+5)
Engines/jit/random_mat_mul_size_10-32           54.9µs ±69%    43.9µs ± 3%   ~     (p=0.841 n=5+5)
Engines/jit/random_mat_mul_size_20-32            162µs ± 2%     161µs ± 2%   ~     (p=1.000 n=5+5)
```