@mathetake 2022-04-12
ok now port the logic to arm64 as well.