@mathetake 2022-04-13
> Is there an overall test to show that compiled code is decoupled from the moduleInstance's pointer?

Well, no it's not possible test, but I just deleted the only embedding of unsafe.Pointer(moduleINstance) https://github.com/tetratelabs/wazero/blob/main/internal/wasm/jit/jit_impl_amd64.go#L3879 here so fine with it.